requirements.txt
absl-py==2.2.2
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==24.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
datasets==3.2.0
decorator==5.2.1
demucs==4.0.1
deta==1.2.0
diffq==0.2.4
dill==0.3.8
distro==1.9.0
dora_search==0.1.12
einops==0.8.1
et_xmlfile==2.0.0
exif==1.6.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2024.9.0
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.5
jiter==0.9.0
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
learntools==0.0.20
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
openai==1.71.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
openpyxl==3.1.5
openunmix==1.3.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plum-py==0.8.7
proglog==0.1.10
propcache==0.2.1
protobuf==4.25.7
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.18.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
safetensors==0.4.5
scipy==1.15.3
sentencepiece==0.2.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.6
submitit==1.5.2
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.47.1
treetable==0.2.5
trl==0.13.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.18.3
